Add logging and stopwatch to track scrapping process
This commit introduces some improvements to the `ScrapemeShopController` class by adding more logging and a stopwatch to track the time taken to scrap the Shop. Additionally, the maximum total number of objects that could be allocated in the WebDriver pool was increased to 20 in 'application.yml'. Guava library is added as a dependency in 'pom.xml' to enable usage of the Stopwatch feature in tracking the scrapping process time. A comment indicating the webpage URL was added to the 'ScrapemeShopPage' class for clarity.