Switch from Chrome to Firefox for web scrapping
The configuration of the web scrapper's environment has been refactored to switch from Chrome to Firefox. This involved making changes to multiple areas, including README instructions, docker compose file, and Java files to accommodate the new browser. Additionally, WebDriver management is no longer automated; instead, the browser is specified in the application.yml file. The reasons for these changes were to ensure better performance and to provide a more flexible configuration, as the desired browser can now be easily switched in the future by modifying a single configuration file.