Refactor web scraping components and configuration
The refactoring includes moving the RequestMapping from class annotation to GetMapping in the ScrapemeShopController. A new concrete class PooledWebDriverPage was introduced for inheritance of the common behaviour. The PageConfig entity was created to handle configuration of the web-scraping pages, which is now defined in the yml file. Adjustments were made to various other classes to adopt these new configurations.