Refactor WebDriver infrastructure (#4)
The previous WebDriver configuration has been refactored to have a better structure and separation of concerns.
The `PoolingWebDriverManager` and `PooledWebDriver` classes are updated to handle timeout values for WebDriver operations. Package naming and file locations have also been re-arranged to be more intuitive and reflective of the functionality they provide.